add business logic create and sellback have to increse and decrese lottery master

+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/lottery/model/dto/UserTicketSummary.java
@@ -1,7 +1,4 @@
 package com.kbtg.bootcamp.posttest.lottery.model.dto;
-
-;
-import jakarta.persistence.Entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -11,7 +8,7 @@ import lombok.Data;
 public class UserTicketSummary {
     private String ticket;
     private String userId;
-    private Long totalTicketAmount;
+    private Integer totalTicketAmount;
     private Double totalTicketBillPrice;
 }
