+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/lottery/model/response/ViewLotteryPurchase.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record ViewLotteryPurchase (
         List<String> tickets,
-        Long count,
+        Integer count,
         Double cost
 ){
 }
